fix(site): create a dedicated comp, remove warns
